requirements.txt
aiogram==3.13.1
gspread==6.1.2
gspread-formatting==1.2.1
google-auth>=2.15.0,<2.42.0
google-generativeai==0.3.2
python-dotenv==1.0.1